Inspect ICMP packets for detailed info
In order to match the ping response or TTL timeout response ICMP packets
to the original ping request, we need to look inside their data section
for the original IP header. We can extract the identification from this
to match it with the original. We store both the original identification
number and the one contained in the response packet in our internal
represenation of the packet.