Look for and store unique intermediate nodes
This simply looks through our stored packets for anything that has the
traceroute source IP as the destination and attempts to find the
original packet that it is a response to. It uses the TTL from the
original packet to determine the number of hops it is away from the
source.