Locate ultimate destination address
This takes into account whether the traceroute program is using UDP or
ICMP for the request messages. For each, the ICMP message type of the
response is different. If traceroute uses ICMP, it's smart and responds
with a ping reply. If it uses UDP, the destination port number is
guaranteed to be invalid and the response will be "port unreachable".