Compute RTT for all intermediate nodes
This currently doesn't take into account the RTT between the source and
the ultimate destination node. The problem with this is that the ICMP
response message does not contain the original IP header like the time
out responses. We need to handle the case for the final node in a unique
way.